Refactored. Moved all the tag processing into it's own file
